Moved init functions.
