+++ dtflw/arguments.py
@@ -95,78 +95,3 @@ def initialize_arguments(argument_class, *values):
         names_and_values = {name: "" for name in values}
 
     return {name: argument_class.create(name, value) for name, value in names_and_values.items()}
-
-
-def init_args(*args):
-    """
-    Initializes widgets using `dbutils.widgets.text` for args:
-    ```
-    flow.notebook(...)
-        .args(...)
-    ```
-
-    Parameters
-    ----------
-    *args
-        Option 1:
-            Given as a variable number of non-keyword arguments, e.g. `args = init_args("a", "b", "c")`.
-            All the args will have an empty string as a value.
-        Option 2:
-            Given as a dictionary with default values, e.g. `args = init_args({'a': 42})`.
-            All non-string values will be stringified.
-
-    Returns
-    -------
-    dict[str: Arg]
-    """
-    return initialize_arguments(Arg, *args)
-
-
-def init_inputs(*inputs):
-    """
-    Initializes widgets using `dbutils.widgets.text` for input tables:
-    ```
-    flow.notebook(...)
-        .input('Orders')
-    ```
-
-    Parameters
-    ----------
-    *inputs
-        Option 1:
-            Given as a variable number of non-keyword arguments, e.g. `inputs = init_inputs("Orders", "Customers")`.
-            All the inputs will have an empty string as a value.
-        Option 2:
-            Given as a dictionary with default values, e.g. `inputs = init_inputs({'Orders': 'orders.parquet'})`.
-            All non-string values will be stringified.
-
-    Returns
-    -------
-    dict[str: Input]
-    """
-    return initialize_arguments(Input, *inputs)
-
-
-def init_outputs(*values):
-    """
-    Initializes widgets using `dbutils.widgets.text` for output tables:
-    ```
-    flow.notebook(...)
-        .output('Orders')
-    ```
-
-    Parameters
-    ----------
-    *args
-        Option 1:
-            Given as a variable number of non-keyword arguments, e.g. `outputs = init_outputs("Orders", "Customers")`.
-            All the args will have an empty string as a value.
-        Option 2:
-            Given as a dictionary with default values, e.g. `outputs = init_outputs({'Orders': 'orders.parquet'})`.
-            All non-string values will be stringified.
-
-    Returns
-    -------
-    dict[str: Output]
-    """
-    return initialize_arguments(Output, *values)
